Fix long restart times in coop play
Some undefined behaviour in NET_SendToAll caused reloads to occasionally
take 5 seconds (the given blocktime) for no reason. This was caused by
checking for net messages on unconnected slots, as the state arrays were
not initialised properly. Applying the same fix here as in other source
ports, see for example
https://github.com/luaman/qforge-1/commit/d4ae231f34ef157f549e181f3159799efffcd645
or the QuakeSpasm source code.