Fix cl_smartjump 1 crashing clients in coop
The check for if the player is in water is done using sv_player, which
is null if only connected as client. It only works for the host.
Instead, cl.inwater can be used, as it is set to the exact condition
waterlevel >= 2 in SV_WriteClientdataToMessage. It should therefore be
equivalent for hosts, and now additionally work properly for clients.