Fix size of buffer for KeepaliveMessage
Since 84d490f6 the NET_MAXMESSAGE is larger than the size of the olddata
array for temporary storage of the message data. This can cause memory
access violations and make clients drop from a server. Happens for
example when a client uses the say command while the server issues a map
restart.

This the size of the array for temporary storage to be NET_MAXMESSAGE
directly. Also move it into static storage, as it became somewhat large
for the stack now (Visual Studio warns about it). Solution is the same
as in QuakeSpasm.