Add cvar to disable attempts to get domain names
On each connection to a server, there is an attempt to find a domain
name for a server from its IP address using a DNS lookup. On Windows,
this can take a long time though. Disabling it can speedup connecting to
servers significantly. As I struggle to see any use case for domain
names, the lookups are disabled by default.