Added QdQStats printed messages
the messages include:
- Playing on X skill
- Exact time was [12:]34.56789
- summed time for marathon runs
Also added check to decide if any kind of qdqstats progs.dat file is loaded. These messages are only shown if there is no qdqstats progs.dat loaded. This way we can traverse the qdqstats feel to any other mods without the need to create additions to several progs.dat in the future. More qdqstats options are expected in the future.
Also fixed the timing in the name of marathon runs recorded by cl_autodemo.