Fix memory access violation in r_viewleaf2
Since 8b4b03a2 there are no oversized arrays with static sizes for vis
data anymore. Instead they are allocated dynamically for the required
size. Unfortunately, in one part that data is iterated over in
increments of 4 bytes, which causes buffer overruns if the dynamically
allocated size is not a multiple of 4. This makes sure to always
allocate a multiple of 4.

Alternatively one could get rid of that iteration in increments of 4
instead i guess. Seems a bit like evil premature optimization to be
honest, so might be the better call.