Fix pause plaque showing when loading a save
This bug was introduced in 3dbeb8ea. Wanted to make it so that a client
who connects to a paused server will get the pause plaque to know what's
going on. Unfortunately the solution in that commit also made it pause
on loading a save, which was not intended of course. To fix that we now
send the pause state as true, only if the pause was actually issued by
the host via Host_Pause_f. Determining that by requiring that cl.paused
and sv.paused are both true instead of only sv.paused.