Fix coop client freeze on server map restart
If a client uses "say" almost in the same instant that the host restarts
the map, that client would often fail to reconnect to the server. The
problem is that the keepalive message is sent using NET_SendMessage
without checking whether a message may be send with NET_CanSendMessage.
So if a keepalive message is sent before the acknowledge message from
the server is received, the datagram system trips up. The sendSequence
and ackSequence counters go out of sync and nothing works anymore.

To fix this, we just skip sending a keepalive message if we cant send a
message. If stuff is loaded faster than receiving the acknowledge
message we probably dont need a keepalive anyways. If a message can be
send in a later call, we can still send a keepalive message then.