Fixed GLSL rendering of luma, caustics and detail textures
Luma textures are applied with GL_ADD instead of GL_DECAL, so for them in GLSL had to use addition instead of mixing by alpha.
Had to completely rewrite generation of texture coords for caustics textures in GLSL.
Also updated the query call of max texture units, raised the available texture units to be a max of 8.