Fixed drawing of caustics textures in GLSL
Sometimes caustics textures were applied on surfaces not underwater. This was caused by the different drawing method in GLSL. We're not copying the 'underwater' surfaces to a separate list and draw them in their own pass (EmitCausticsPolys), but draw them all in one pass. This means we need to call glDrawElements per waterline to properly separate caustics/detail surfaces when applying the appropriate textures.
Also optimized GLSL code a tiny bit when drawing caustics/detailed textures.