Fixed vertex lighting in GLSL
Had to learn (google) about uniforms and buffers in GLSL, but finally came up with a solution on transfering the 65k bytes 'vlighttable' to the shader using a texture buffer.
Also made use of a uniform buffer as a bit of an optimization, might be beneficial for other uniform fields in the future.

Small change in r_world_program: turbsin is not calculated anymore, but found out it is possible to simply define the const array in GLSL.