Changed default values of gl_externaltextures_* cvars to be OFF
Most of the players like the old vanilla look, so I decided to not force the hi-res images by default. In the next version, I'm planning another menu overhaul so that switching external textures ON would be more intuitive.
Also changed that gl_externaltexture_gfx is now editable from the menu.