Console background and font is reloaded when gl_externaltextures_gfx gets changed
Also, if a hi-res version of the console background is available, it will only be used if gl_externaltextures_gfx is 1.
Fixed a small glitch when GL_LINEAR filter wasn't applied to the console font in case of gl_smoothfont 1, by a side effect of GL_NEAREST texturemode - when it was used.