Fixed bug when skybox images were reloaded unnecessarily all the time
Happened on map ad_heresp2 which toggles between 2 skyboxes by teleporting between two areas.
Added a check for the name of the currently loaded skybox to prevent reloading it again and again all the time.
Also changed the code to store the already loaded skybox images in the cache, so the game does not try to reload such skyboxes that were loaded previously.