BSP textures are identified more clearly (in memory)
There was a problem in ad_necrokeep, where a bmodel had a different texture with the same name and this texture overwrote the original one from the map itself - since the in-memory identifier only used the texture name itself.
I've extended the identifier to have the BSP name as a prefix as well, to make it clear which texture relates to which bsp file.