Fixed rare issue when rendering water surfaces was bugged
Seen only on jam2_lunaran, there was a separate brush model drawn having SURF_DRAWLAVA flag set. Since it is not part of the map bsp file, it was not treated as water surface and had to be fixed in R_DrawBrushModel.
Also fixed the issue when changing r_wateralpha wasn't have any effect on the map_wateralpha/lavaalpha/slimealpha variables set while map load.