Vertex lighting is not computed with GLSL on ATI/AMD gpus
Couldn't find the exact reason why, but cannot use the texture buffer feature with GLSL on ATI/AMD cards to store and reload the vertex lighting information.
So until I come up with a solution, the most simple way is to ignore the texelFetch() GLSL call when computing vertex lighting data.