don't append a record if no update this packet
There are initial packets that don't include the view entity.  In this
case the zero-initialized record is used.  This mostly has an effect for
demos in which do not start at the start of the game (ie. `record` was
run in the middle of a game).  In this case, without the fix, the ghost
can be seen slowly interpolating from the origin to the its real
starting point.