don't show delta time if too far from ghost's path
This helps with a couple of cases:

a)  The ghost never finished, so the delta match with a much different
point.

b)  The ghost took a different route, in which case the delta doesn't
make much sense anyway.

I chose a range of 256 units here, which seems sensible.  I am open to
tweaking it.