Found a better way to fix intermission cam
The previous fix (from MarkV) resulted broken svc_cutscene cameras.
The root cause of the misaligned intermission cams (during demoplayback) is that the last svc_setangle received just before the intermission is not transferred to the player entity, which angles are used in V_CalcIntermissionRefdef. So I figured out I can save the vector of this svc_setangle to a variable and directly set it in svc_intermission.