Changed show_speed calculation to be more accurate
In Quake, an entity's velocity is sent to the client as 3 bytes, to save network bandwidth. This unfortunately results arithmetically incorrect velocity vector values after resizing them to float.
To overcome this issue, I store the velocity vector's float values to a global variable, to use them only when printing the player's actual speed.