Using overbright on brush and alias models as in QS
Using a separate palette with overbright colors resulted blown out textures on alias models at certain lighting conditions. It also applied overbrights on the stock, palette based texture colors, but not on external textures.
The solution in QS is more clean and gives more freedom for the user as well with the gl_overbright and gl_overbright_models cvars.
The only disadvantage is the loss of the 10bit color buffer implementation for lighting - in case of rendering without the use of GLSL scripts.