Clamping physframetime to 0.1 at max
When a map load takes longer than 0.1 sec, physframetime could grow beyond 0.1 which resulted a missing server frame at level starts.
At first this was thought to be a serious issue as it resulted very early start times (1.31x sec) but it turned out with today's PCs this is actually normal.
Thanks to @matthewearl for the contribution ❤️