Created separate VS project file for the unused software build
I was tired from the shitload of unused configurations like the software debug and release ones so I decided to move those to a separate project called JoeQuake-sw. It's not compiling now (as it is, for ages) and I also omitted it from building. I just left it there if someone in the future wants to tinker with it.
Also removed useless debug.h and mgraph.h headers. These must come from the external /scitech subfolder. Will commit these missing subfolders in the future.