unpause on demoskip
The old behaviour has no apparent effect when paused --- after skipping
the view only changes when unpause is pressed.  I think this is
confusing since it looks like the skip didn't happen at all unless the
player realises they must unpause.  Ideally the first frame of the new
level would load but the game would be paused from that point.  This
seems tricky to implement though, so I am simply unpausing on skip as a
compromise