Ghost now properly works on levels ending with svc_cutscene
The fix is to handle svc_cutscene events as svc_intermission (or finale) BUT only on relevant maps where cutscene messages are sent when the level is finished. Unfortunately this cannot be determined, so can only use our experience. So far only hipend and mexx8 levels are discovered to be planned this way. The list of supported levels should be extended in the future.