maximise number of marathon splits
At the end of a map, show as many ghost split times as possible. A map
is shown on the intermission if:

i) it is one of the maps the player has completed in this session (ie.
   without an intervening `map` or `load`).
ii) there is a level in the ghost demo on the same map (in this case the
    first such map in the ghost demo is used as the comparison time).
iii) All later maps the player completed (if any) are also visible.

The old logic would only show ghost splits if all previous player maps
had a ghost map*.  I think this way is useful, for example, to compare
ghost episode marathons with full game demos.

*the intermission screen splits (rather than the console splits) would
also sometimes show wrong splits when there are no ghost splits
printed to the console splits, this is also fixed in this commit.