fix ghost when rewinding through finale
When rewinding through an intermission, `cls.marathon_time` and
`cls.marathon_level` need decreasing otherwise when the demo is played
forwards again the level will end up being counted twice, which results
in two ghost splits being shown for the current level.

Currently this is handled by `PrintFinishTime` which decreases the
`cls.marathon*` variables if the demo is rewinding.  This works fine for
plain `svc_intermission`, however with `svc_finale` and `svc_cutscene`
the call to `PrintFinishTime` is guarded by `!cl.intermission`, which is
always false when the demo is reversing.

This commit changes the logic to decreasing the marathon variables
whenever cl.intermission goes from being non-zero to zero.