fix coop marathon tracking
In co-op it is not possible for a non-host to determine whether a
map change is a changelevel or otherwise*.  Since a map change
changelevel is used to trigger a marathon reset, this means non-host
clients cannot track marathons correctly.  Nevertheless, the ghost code
was reading `cls.marathon_level` which could cause strange effects in
co-op situations.

The commit fixes this by replacing `cls.marathon_level` with
`gmi->num_levels`, which is equal to `cls.marathon_level` unless
connected as a non-host client.  In the non-host case `gmi->num_levels`
is 0 if not on an intermission, and 1 otherwise.

This commit also incidentally fixes another issue with
`ghost_marathon_split`, which was no longer being reset to zero after
running `restart` (thanks to Muty for reporting).

* Is this true?  Maybe non-hosts are always booted unless a changelevel
  occurs?