Workaround for skyfog value not being reset
When mods do not reset r_skyfog to a default value on the start of a
map, instead it will use the value from the end of the previous map.

This happens for example with ad_tears: The final fight area of ad_tears
sets r_skyfog to a very low value, so there is basically no fog. After
ending the map, you return back to the start map. But it does not reset
the fog to its default value of 0.5, so you get a wrong looking sky
without any fog. This should probably be fixed in Arcane Dimensions
itself?

This is an engine-side workaround for this. Instead of using r_skyfog
which may be modified by mods, we use a new seperate variable
r_skyfog_default to initialize the fog when loading a map. Once a mod
does any changes to r_skyfog, that value will be used instead.