Fix speedometer in multiplayer
The existing check for when we have to fallback to the inaccurate client
speed is flawed.

Checking the coop cvar makes sure that the host falls back to showing
client-side speed when they are hosting a coop server. But in other
multiplayer scenarios like deathmatch, their speedometer will be broken.
Also when just playing singleplayer but coincidentally having set the
coop cvar to 1, they will get the less accurate speed display for no
good reason.

Non-host clients of multiplayer games on the other hand will have no
speed display at all, unless their unused coop cvar happens to be 1 by
chance.

We really only need to choose the fallback method if either we are
not hosting and just connected to a server (!sv.active) or if we are
hosting a game with multiple players (sv.maxplayers != 1).