Allow overriding spawnparams
Extends the spawnparam features added in #41 to be used for singleplayer
segmented runs that need start stats when starting levels with `map` or
`record` commands. Without the overriding option, those features only
helped when starting runs with `restart` like in coop.

Basically an engine-side variant of what qdqstats provides, so that we
can work with custom mods too.