Add support for LiveSplit autosplitter
This will allow the LiveSplit autosplitter to automatically find all
the memory addresses. It scans for the magic data string that is
directly followed by a list of all the data that it needs to run the
autosplitting. That means any new JoeQuake version will immediately
have autosplitting support, without any manual work.